Relax Python version to "python3"
Python 3.7 was not installed on my machine and I believe it might
be fine to just say "python3" here. If something fails on that,
people using this will probably have to go to "pyenv" anyway.

In order to add back an assertion for Python>=3.7, it brings in
a corresponding runtime check.