Add a version warning when Crate Docs is out of date
When running `make`, a request is made to GitHub to determine the latest Crate
Docs release. The latest version is compared to the pinned build version and if
there is mismatch, a warning is issued.

I spent way too long trying to get this working with Make's inbuilt `$(info
...)` function because that is the idiomatic way of issuing these kind of
notices. For some reason, Make absolutely refused to properly execute shell
code for the purpose of display output. Eventually I gave up and implemented
the display output with a PHONY target. This has the benefit of allowing us to
use ANSI escapes, which is not possible with `$(info ...)`.

Closes https://github.com/crate/crate-docs/issues/21