Rename `demo-docs` back to `docs`
I ran into an issue upgrading crate-admin to the 1.1.0 release. The standard
procedure (documented in `demo-docs/Makefile`) is to copy the
`demo-docs/Makefile` to your own project.

Unfortunately, because this Makefile defines the `DOCS_DIR` variable to
`demo-docs`, if you copy this Makefile to your own project (using `docs` as
`DOCS_DIR` as is convention), you will encounter the following error when
running any of the core `make` commands:

    ERROR: Could not open requirements file: [Errno 2] No such file or directory: 'demo-docs/requirements.txt'

We could work around this in the same way we work around the Git checkout of
the common-build. However, given that the demo docs serves the purpose of being
a reference implementation, I actually think it would be more correct to rename
it back to `docs`, as is convention.

I still think the renaming of `src` to `common-build` is a good choice. But
there is no real reason to have `demo-docs` as opposed to `docs`, and indeed,
we might even want to expand the scope of the demo docs project to actually
include real documentation of the common build system.