Fix the use of `--keep-going` option
The `--keep-going` option does not work with `sphinx-autobuild`, resulting in
an error when you run `make dev`.

To fix this issue, I moved `--keep-going` to a new variable,
`SPHINX_OPTS_CI`. The `_CI` suffix indicates that this option is for use
primarily by CI tools (i.e., `make html` and `make check`).