Fix CHANGES.rst to add a note about breaking changes
The 1.1.0 release should have included a note about breaking changes. This
changeset adds that to the change log. The GitHub tag description has already
been updated.

For SemVer reasons, we should have released 2.0.0 instead of 1.1.0. We could
just ignore this, or we could ship this recent batch of fixes as 2.0.1 and make
a note about the reason why we skipped 2.0.0.