Add the `outline` program for outlining RST-style document headings
I am currently working on <https://github.com/crate/crate/pull/11277>, and
after I converted `crate.yml` to use RST-style heading syntax, I immediately
noticed a problem. The outline of the file looked like this:

    Basic setup
      Memory
      Persistence
      Networking
      Clustering
    Advanced settings
      JMX monitoring plugin
      Authentication
        Host-based access
        Trust-based authentication
      User-defined functions (UDFs)
      Secured communications (SSL/TLS)
        Optional configuration for the truststore
      Cluster settings
        Graceful stop
        Routing allocation
      Node settings
      Paths
      Memory
      Network and HTTP
      Metadata gateway
      Recovery throttling
      Discovery settings
      Sharding
      Field-data circuit breaker
      Threadpools
      Metadata
      Garbage collection (GC) logging
      SQL
      Query circuit breaker
      Usage-data collector (UDC)
      Backup and restore (aka snapshots)
      PostgreSQL wire transport

The "Basic setup" section is okay because it does not attempt to mirror
anything in the documentation. However, the "Advanced settings" section is
supposed to list all CrateDB configuration options (it does not list all
options, but that is potentially another topic for another day).

The most immediate problem with this outline is that the CrateDB documentation
splits the configuration options into two major sections:

- Node-specific settings
  <https://crate.io/docs/crate/reference/en/latest/config/node.html>

- Cluster-wide settings
  <https://crate.io/docs/crate/reference/en/latest/config/cluster.html>

In the `crate.yml` file, we have:

- A "Cluster settings" section with two subsections: "Graceful stop" and
  "Routing allocation"

- A "Node settings" section with no subsections

These two sections appear in reverse order and are positioned halfway down the
file. The rest of the sections resemble a seemingly random assortment of
subsections from both the "Node-specific settings" and "Cluster-wide settings"
documents.

The result is a confusing and misleading mess of information. However, the
cause of this problem is straightforward enough: The two "sources of truth"
(re what the config options are) have had over half a decade to drift apart.

For comparison, below are the outlines of the two sections in the docs.

Node-specific settings:

    Basics
    Node types
    Read-only node
    Hosts
    Ports
    Paths
    Plug-ins
    CPU
    Memory
    Garbage collection
    Authentication
      Trust authentication
      Host-based authentication
          HBA entries
    Secured communications (SSL/TLS)
    Cross-origin resource sharing (CORS)
    Blobs
    Repositories
    Queries
    JavaScript language
    Custom attributes

Cluster-wide settings:

    Non-runtime cluster-wide settings
    Collecting stats
    Shard limits
    Usage-Data Collector
    Graceful stop
    Bulk operations
    Discovery
      Unicast host discovery
        Discovery via DNS
        Discovery on Amazon EC2
        Discovery on Microsoft Azure
    Routing allocation
      Awareness
      Balanced shards
      Cluster-wide allocation filtering
      Disk-based shard allocation
    Recovery
    Memory management
    Circuit breakers
      Query circuit breaker
      Field-data circuit breaker
      Request circuit breaker
      Accounting circuit breaker
      Stats circuit breakers
      Total circuit breaker
    Thread pools
      Settings for fixed thread pools
    Metadata
      Metadata gateway

As I started to attempt to bring `crate.yml` in line with the documentation, I
realized it would be difficult to do this by hand. I also realized that
the two sources of truth would quickly drift apart again. An unfortunate
consequence of this practically unavoidable violation of the principle of Do
Not Repeat Yourself.

I have used `diff` to verify the contents of tarballs before, and I realized
that I could use the same method to compare the two sources of truth if only I
could generate outlines of the headings.

I wrote the `outline` script added by this commit to satisfy that wish. Read
the docstring in the file for details (there is an example showing how the
program can be used with `diff` to satisfy my use-case).

By way of demonstration, I generated the outlines in this commit message
(above) using the `outline` program.

My original plan was to commit this file to the primary CrateDB repository,
with the thought that we might eventually use it for automated
testing. However, I realized that a secondary use for this program is as a
validation tool for our RST style guide more generally.

For example, when I first ran `outline` on `cluster.rst` in the docs, I noticed
that some sections were at the wrong level (e.g., one level too deep). The
standard RST parser can only tell you if your use of underlines is inconsistent
within one document. It cannot tell you if your use of underlines is
inconsistent between documents.