Add the `check-crate-conf` command
This script uses the `outline` program to compare the "node" and "cluster"
sections in the stock `crate.yml` configuration file with the respective
`node.rst` and `cluster.rst` files in the CrateDB documentation.

Ideally, the structure and content of the stock configuration file match the
documentation's structure and content. Unfortunately, however, because we have
two "sources of truth", they have significantly fallen out of sync.

You can use this script to:

- Periodically check the alignment of both source of truth, and
- Help you reconcile any differences you find

See the source code for extended documentation.